requirements.txt
pynput>=1.7.6
pyperclip>=1.8.2
python-dotenv>=1.0.0
